\name{hyper2-package}
\alias{hyper2-package}
\docType{package}
\title{
\packageTitle{hyper2}
}
\description{
\packageDescription{hyper2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hyper2}
\packageIndices{hyper2}

  A generalization of the Dirichlet distribution, using a more
  computationally efficient method than the \pkg{hyperdirichlet}
  package.  The software is designed for the analysis of order
  statistics and team games.

  There are several vignettes that come with the package: see
  \href{https://robinhankin.github.io/hyper2/}{the io readme} (under
  \sQuote{articles}), or listed by \code{vignette(package = "hyper2")}.

  The hex sticker features the \dQuote{draw monster}, a concept
  introduced in Hankin (2020) as a reified entity, representing the
  winner of a drawn chess game.  The character was visualized and
  sketched by my daughter Annabel.  In the chess world, draws are often
  seen as dull and undesirable, and the draw monster embodies this
  sentiment: he is a dampening presence, a metaphorical wet blanket.
  The drawing conveys this lack of fighting spirit well, capturing the
  essence of an uninspired and anticlimactic outcome.
  
}
\author{
\packageAuthor{hyper2}

Maintainer: \packageMaintainer{hyper2}
}
\references{
  \itemize{
    \item 
    R. K. S. Hankin (2010). \dQuote{A Generalization of the Dirichlet Distribution},
    \emph{Journal of Statistical Software}, 33(11), 1-18,
    \doi{10.18637/jss.v033.i11}
    \item R. K. S. Hankin 2017. \dQuote{Partial Rank Data with the hyper2
      Package: Likelihood Functions for Generalized Bradley-Terry Models}.
    \emph{The \R Journal} 9:2, pages 429-439, \doi{10.32614/rj-2017-061}
    \item R. K. S. Hankin 2024.  \dQuote{Generalized Plackett-Luce
      Likelihoods}, \emph{Journal of Statistical Software}, 109(8), 1-17,
    \doi{10.18637/jss.v109.i08}
  }
}
\keyword{ package }
\examples{
icons
maxp(icons)
}
