\name{motoGP}
\alias{motoGP}
\alias{motoGP_2019}
\alias{motoGP_2019_table}
\docType{data}
\title{MotoGP dataset}
\description{Race results from the 2019 Grand Prix motorcycling season}
\usage{data(motoGP_2019)}
\format{A \code{hyper2} object that gives a likelihood function}
\details{

Object \code{motoGP_2019} is a \code{hyper2} object that gives a
likelihood function for the strengths of the competitors of the 2019
Grand Prix motorcycling season.  It is created by
\code{ranktable2supp(motoGP_2019_table)}. 
}
\references{

Wikipedia contributors. (2020, February 8). 2019 MotoGP season. In
\emph{Wikipedia, The Free Encyclopedia.} Retrieved 08:16, February 20,
2020, from \url{https://en.wikipedia.org/w/index.php?title=2019_MotoGP_season&oldid=939711064}
}
\seealso{\code{\link{ranktable2supp}}}
\examples{
data(motoGP)
}
