\name{ranktable_to_ordertable}
\alias{ranktable_to_ordertable}
\title{Convert rank tables to order tables}
\description{
  Convert rank tables (as generated by \code{rrank()},  for example) to
  order tables like the formula 1 tables.
}
\usage{
ranktable_to_ordertable(x)
}
\arguments{
  \item{x}{Rank table, as produced by \code{rrank()}
  }
}
\details{The function is trivial.}
\value{A matrix}
\author{Robin K. S. Hankin}
\seealso{
  \code{\link{rrank}},
  \code{\link{ordertable2supp}}
}
\examples{

p <- (5:1)/15
names(p) <- letters[1:5]
maxp(ordertable2supp(ranktable_to_ordertable(rrank(12,p,rnames=month.abb))))
}
