\name{surfing}
\alias{surfing}
\alias{surfing_table}
\alias{surfing_maxp}
\alias{surfing_venuetypes}
\docType{data}
\title{Surfing dataset}
\description{Data from the 2019 World Surf League (WSL) tour}
\usage{data(surfing)}
\details{

  The package contains four datasets from WSL 2019:

  \itemize{
    \item \code{surfing}, a log likelihood function for the strengths of
    the competitors
    \item \code{surfing_maxp}, corresponding precalculated evaluate
    \item \code{surfing_venuetypes}, a dataframe showing the beach types
    at the different venues of the tour
  }

\generatedbynolink{inst/surfing.Rmd}{surfing.rda}

}
\references{

T. C. Driver and R. K. S. Hankin 2023.  \dQuote{Analysis of competitive
surfing tournaments with generalized {B}radley-{T}erry likelihoods},
\emph{Journal of Sports Analytics}, 9:133--140.
\doi{10.3233/JSA-220596}
}

  
\author{Robin K. S. Hankin}
\examples{
dotchart(surfing_maxp)

}
\keyword{datasets}

