\name{attemptstable2supp3}
\alias{attemptstable2supp}
\alias{attemptstable2supp3}
\title{Translate attempt tables to hyper3 support functions}
\description{\loadmathjax
description here
}
\usage{
attemptstable2supp3(a, decreasing, give.supp=TRUE, dnf.last=TRUE)
}
\arguments{
  \item{a}{Data frame, see details}
  \item{decreasing}{Boolean, with \code{TRUE} meaning that the highest
      score wins [e.g. javelin distances] and \code{FALSE} meaning that the
      lowest score wins [e.g. times for a race]}
  \item{give.supp}{Boolean, return the support function or the order statistic}
  \item{dnf.last}{Boolean, should NA entries count as coming last (TRUE)
or  be ignored (FALSE)}
}
\details{

Function \code{attemptstable2supp3()} is intended for use on attempts
tables like \code{javelin}.

These objects can be generated by running script
\code{inst/javelin.Rmd}, which includes some further discussion and
technical documentation, and creates file \code{javelin.rda} which
resides in the \code{data/} directory.

}

\value{Returns a \code{hyper3} object}
\author{Robin K. S. Hankin}
\seealso{\code{\link{ordertable2supp}},\code{\link{javelin}}}
\examples{

jj <- javelin_table[1:3,]
jj
attemptstable2supp3(jj)

}
