\name{carcinoma}
\alias{carcinoma}
\alias{carcinoma_table}
\alias{carcinoma_count}
\docType{data}
\title{Carcinoma dataset discussed by Agresti}
\description{A dataset considered by Agresti.  Seven clinicians are
  asked whether they see evidence for carcinoma on different patients.}
\usage{data(carcinoma)}
\format{A \code{hyper2} object that gives a likelihood function}
\details{

These objects can be generated by running script \code{inst/carcinoma.Rmd},
which includes some further discussion and technical documentation, and
creates file \code{carcinoma.rda} which resides in the \code{data/}
directory.

}
\references{

A. Agresti, 2002.  "Categorical data analysis".  John Wiley and
Sons. Table 13.1, p542.
 }
\seealso{\code{\link{race3}}}
\examples{
data(carcinoma)
ordertable_to_ranktable(carcinoma_table)
pie(carcinoma_maxp)
}

