\name{curling}
\alias{curling_table}
\alias{curling}
\alias{curling1}
\alias{curling2}
\alias{curling_maxp}
\alias{curling1_maxp}
\alias{curling2_maxp}
\docType{data}
\title{Curling at the Winter Olympics, 1998-2018}
\description{
Data for women's Olympic Curling  at the 2002 Winter Olympics.
}
\usage{data(curling)}
\details{

  There are five datasets loaded by \code{data("curling")}:

  \itemize{

    \item \code{curling_table}, an order table for Winter Olympics years
    1998,2002,2006,2010,2014, and 2018 for 13 countries.

    \item \code{curling1}, a log likelihood function on the assumption
    that not attending (indicated by \code{NA}) is equivalent to a
    \code{DNS} in Formula 1

    \item \code{curling2}, a log likelihood function on the assumption
    that not attending is noninformative

    \item \code{curling1_maxp} and \code{curling2_maxp}, corresponding
    evaluates
  }


These objects can be generated by running script
\code{inst/curling.Rmd}, which includes some further discussion and
technical documentation and creates file \code{curling.rda} which
resides in the \code{data/} directory.

}
\references{
  \itemize{
    \item

Wikipedia contributors. Curling at the Winter Olympics [Internet].
Wikipedia, The Free Encyclopedia; 2021 Jan 7, 14:23 UTC [cited 2021 Jan
21]. Available from:
\url{https://en.wikipedia.org/w/index.php?title=Curling_at_the_Winter_Olympics&oldid=998891075}
  }
}
\author{Robin K. S. Hankin}
\examples{
data(curling)
dotchart(curling1_maxp)

}
\keyword{datasets}

