\name{hepatitis}
\alias{hepatitis}
\alias{hepatitis_table}
\alias{hepatitis_maxp}
\alias{hepatitis_count}
\docType{data}
\title{Hepatitis dataset discussed by Agresti}
\description{A dataset considered by Agresti}
\usage{data(hepatitis)}
\format{A \code{hyper2} object that gives a likelihood function}
\details{

Object \code{hepatitis_table} is drawn from Agresti, table 12.16, page
533.  Object \code{hepatitis} is a likelihood function of class
\code{lsl} and \code{hepatitis_maxp} a pre-calculated evaluate.

These objects can be generated by running script \code{inst/hepatitis.Rmd},
which includes some further discussion and technical documentation, and
creates file \code{hepatitis.rda} which resides in the \code{data/}
directory.

}
\references{
A. Agresti, 2002.  "Categorical data analysis".  John Wiley and
Sons. Table 13.1, p542.
 }
\seealso{\code{\link{race3}},\code{\link{hepatitis}}}
\examples{
pie(hepatitis_maxp)
}

