\name{javelin}
\alias{javelin}
\alias{javelin_maxp}
\alias{javelin1_maxp}
\alias{javelin2_maxp}
\alias{javelin_table}
\docType{data}
\title{Javelin dataset}
\description{Olympic javelin results}
\usage{data(javelin)}
\format{A \code{hyper2} object that gives a likelihood function}
\details{

See \code{inst/javelin.Rmd} for details

}
\seealso{\code{\link{attempttable2supp3}}}
\examples{
pie(javelin1_maxp)
}

