\name{javelin}
\alias{javelin}
\alias{javelin1}
\alias{javelin2}
\alias{javelin_vector}
\alias{javelin_maxp}
\alias{javelin1_maxp}
\alias{javelin2_maxp}
\alias{javelin_table}
\docType{data}
\title{Javelin dataset}
\description{

  An \code{attemptstable} object showin results from the men's javelin,
  2020 Summer Olympics.
  
}
\usage{data(javelin)}
\details{

  \itemize{
    \item \code{javelin_table}, an \code{attemptstable} object detailing
    the throw distances of eight competitors (diacritics have been
    removed) for each of six throws
    \item \code{javelin1} and \code{javelin2}, support functions
    corresponding to the weighted Plackett-Luce likelihood.  The suffix
    \dQuote{1} means that no-throws are counted as losing attempts;
    suffix \dQuote{2} means that no-throws are ignored.
    \item \code{javelin1_maxp} and \code{javelin2_maxp} are the
    corresponding maximum likelihood estimates for the players'
    strengths
    \item{javelin_vector} is a named vector with elements being the
    throw distances and names being the thrower
  }

These objects can be generated by running script
\code{inst/javelin.Rmd}, which includes some further discussion and
technical documentation, and creates file \code{javelin.rda} which
resides in the \code{data/} directory.

}
\author{Robin K. S. Hankin}
\note{

The format for Olympic javelin results is common to six of the field
events: long jump, triple jump, shot put, discus throw, hammer throw,
and javelin (high jump and long jump have a different format).  The
\code{inst/} directory on github includes some other datasets such as
\code{olympics_2020_womens_shotput.txt} in the same form as
\code{inst/javelin.txt}.

  }
\seealso{\code{\link{attemptstable2supp3}}}
\examples{

attemptstable2supp3(javelin_table)

pie(javelin1_maxp)
}

