\name{javelin}
\alias{javelin}
\alias{javelin1}
\alias{javelin2}
\alias{javelin_vector}
\alias{javelin_maxp}
\alias{javelin1_maxp}
\alias{javelin2_maxp}
\alias{javelin_table}
\docType{data}
\title{Javelin dataset}
\description{

  Results from the men's javelin, 2020 Summer Olympics.
  
  \itemize{
    \item \code{javelin_table}, a dataframe in the form of an
    \dQuote{attempts table}, detailing the throw distances of eight
    competitors (diacritics have been removed) for each of six throws
    \item \code{javelin1} and \code{javelin2} Support functions
    corresponding to the weighted Plackett-Luce likelihood.  The suffix
    \dQuote{1} means that no-throws are counted as losing attempts;
    suffix \dQuote{2} means that no-throws are ignored.
    \item \code{javelin1_maxp} and \code{javelin2_maxp} are the
    corresponding maximum likelihood estimates for the players'
    strengths
    \item{javelin_vector} is a named vector with elements being the
    throw distances and names being the thrower
  }
}
\usage{data(javelin)}
\format{As detailed above}
\details{

These objects can be generated by running script
\code{inst/javelin.Rmd}, which includes some further discussion and
technical documentation, and creates file \code{javelin.rda} which
resides in the \code{data/} directory.

}
\seealso{\code{\link{attemptstable2supp3}}}
\examples{
pie(javelin1_maxp)
}

