\name{karate}
\alias{karate}
\alias{karate_table}
\alias{karate_maxp}
\docType{data}
\title{Karate dataset}
\description{
  Dataset from the 2018 World Karate Championships, men's 67kg.
}
\usage{data(karate)}

\details{

Object \code{karate_table} is a dataframe of results showing results
from the 2018 World Karate Championships, men's 67kg; \code{karate} is
the associated likelihood function and \code{karate_maxp} the evaluate.

These objects can be generated by running script \code{inst/karate.Rmd},
which includes some further discussion and technical documentation and
creates file \code{karate.rda} which resides in the \code{data/}
directory.

}
\references{
\url{https://en.wikipedia.org/wiki/2018_World_Karate_Championships}
}

\note{

Table \code{karate_table} misses uninformative matches, that is,
competitions with 0-0 results.

}
\seealso{\code{\link{zapweak}}}
\examples{
summary(karate)
}
