\name{ranktable_to_ordertable}
\alias{ranktable_to_ordertable}
\alias{ordertable_to_ranktable}
\title{Convert rank tables to and from order tables}
\description{
  Convert rank tables (as generated by \code{rrank()},  for example) to
  order tables like the formula 1 tables.
}
\usage{
ranktable_to_ordertable(xrank)
ordertable_to_ranktable(xorder)
}
\arguments{
  \item{xrank}{Rank table, as produced by \code{rrank()}}
  \item{xorder}{Order table}
}
\details{
Function \code{ranktable_to_ordertable()} is trivial;
\code{ordertable_to_ranktable()} less so.  The prototype for order
tables would be \code{skating_table}. 
}
\value{An order table or rank table}
\author{Robin K. S. Hankin}
\seealso{
  \code{\link{rrank}},
  \code{\link{ordertable2supp}}
}
\examples{

p <- (5:1)/15
names(p) <- letters[1:5]
xrank <- rrank(12,p,rnames=month.abb)

ranktable_to_ordertable(xrank)
xorder <- ordertable_to_ranktable(ranktable_to_ordertable(xrank))

identical(xrank,ordertable_to_ranktable(ranktable_to_ordertable(xrank)))

maxp(ordertable2supp(xorder))  # should be close to p

ordertable_to_ranktable(skating_table)
}
