\name{RCLF}
\alias{RCLF}
\alias{RCLF3}
\alias{RCLF3_max}
\alias{RCLF3_max_lambda}
\alias{Rangers}
\alias{Celtic}
\alias{Livingston}
\alias{Falkirk}
\docType{data}
\title{Dataset from four Scottish football clubs}
\description{

  These objects refer to results from football matches among four
  Scottish clubs: Rangers, Celtic, Livinston, Falkirk.
  


}
\usage{data(RCLF)}
\details{
The objects are created by \code{inst/home_advantage.Rmd}.
}
\source{
  Data obtained from \code{www.worldfootball.net}
}
\author{Robin K. S. Hankin}
\examples{
data(RCLF)
RCLF_table
RCLF3_max
RCLF_lambda_max
}
\keyword{datasets}
