\name{summary.hyper2}
\alias{summary.hyper2}
\alias{print.summary.hyper2}
\title{Summary method for hyper2 objects}
\description{Give a summary of a hyper2 object, and a print method}
\usage{
\method{summary}{hyper2}(object, ...)
\method{print}{summary.hyper2}(x, ...)
}
\arguments{
  \item{object,x}{Object of class hyper2}
  \item{...}{Further arguments, currently ignored}
}
\details{
Mostly self-explanatory, based on the equivalent in the \code{untb} package.
}
\author{Robin K. S. Hankin}
\seealso{\code{\link{hyper2}}}
\examples{
summary(icons)
}
