\name{sushi}
\alias{sushi}
\alias{sushi_table}
\alias{sushi_maxp}
\docType{data}
\title{Sushi dataset}
\description{
  Dataset from the \emph{preflib} website, discussing sushi preferences.
}
\usage{data(sushi)}

\details{

Sushi preference dataset.

These objects can be generated by running script \code{inst/sushi.Rmd},
which includes some further discussion and technical documentation and
creates file \code{sushi.rda} which resides in the \code{data/}
directory.
}

\examples{
sushi
sushi_maxp
}
