\name{sushi}
\alias{sushi}
\alias{sushi_table}
\alias{sushi_maxp}
\docType{data}
\title{Sushi dataset}
\description{
  Dataset from the \emph{preflib} website, discussing sushi preferences.
}
\usage{data(sushi)}

\details{

Object \code{sushi_table} is a dataframe of results showing results
from the 2018 World Sushi Championships, men's 67kg; \code{sushi} is
the associated likelihood function.  There are two maximum likelihood
estimates given; \code{sushi_maxp}, the evaluate as returned by
\code{maxp()}, and \code{sushi_maxp}, returned by \code{zermelo()} [the
value given by \code{maxp()} itself is less likely].

These objects can be generated by running script \code{inst/sushi.Rmd},
which includes some further discussion and technical documentation and
creates file \code{sushi.rda} which resides in the \code{data/}
directory.
}
\references{
\url{https://en.wikipedia.org/wiki/2018_World_Sushi_Championships}
}

\note{

Table \code{sushi_table} misses uninformative matches, that is,
competitions with 0-0 results.

}
\seealso{\code{\link{zapweak}}}
\examples{
summary(sushi)
}
