\name{universities}
\alias{universities}
\docType{data}
\title{New Zealand University ranking data}
\description{Times Higher Education World University Rankings}
\usage{data(universities)}
\format{A \code{hyper2} object that gives a likelihood function for
  ranking of NZ universities}
\details{


  \url{https://www.timeshighereducation.com/world-university-rankings/2020/world-ranking#!/page/0/length/25/locations/NZ/sort_by/rank/sort_order/asc/cols/stats}


}
\seealso{\code{\link{ordertable}}}

